refactor: update author and layout templates
This commit updates the author and layout templates to include the
last updated date. This change enhances the transparency of content
updates and provides users with more context about the recency of the
information presented.

It also includes minor code improvements in the utils.py file to ensure
better handling of navigation elements. It also fixes a minor issue
with aria-controls attribute assignment and copy URL button is now
working.

Signed-off-by: Akshay Mestry <xa@mes3.dev>