fix(read time): correct estimated read time
This commit fixes the issue with the estimated read time calculation on
articles. The previous implementation did not account for images and
code blocks, leading to inaccurate read time estimates. The new logic
includes a more comprehensive word count that factors in these
elements, providing users with a better understanding of the time
required to read the content.

This also fixes the border and removes the use of box-shadow for code
blocks and tab labels, aligning with the overall design aesthetics of
the website.

Signed-off-by: Akshay Mestry <xa@mes3.dev>